xwlb-crawler/requirements.txt
# 新闻联播爬虫系统 - Python依赖
# 安装: pip install -r requirements.txt

# 核心爬取
playwright>=1.40.0
playwright-stealth>=1.0.0
requests>=2.31.0
fake-useragent>=1.4.0

# NLP分析
jieba>=0.42.1
gensim>=4.3.0
snownlp>=0.12.3

# 定时调度
apscheduler>=3.10.0

# 数据存储
sqlalchemy>=2.0.0

# 配置与工具
pydantic>=2.5.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# 日志
loguru>=0.7.0

# 开发工具（可选）
# pytest>=7.4.0
# black>=23.0.0
